add SearchResultsPage, changed search test accordingly to return searchresultspage
